\name{PONG2-internal}
\alias{hlaAllele}
\alias{hlaAlleleDigit}
\alias{hlaAlleleSubset}
\alias{hlaAttrBagging}
\alias{hlaBED2Geno}
\alias{hlaCheckSNPs}
\alias{hlaClose}
\alias{hlaCombineAllele}
\alias{hlaCombineModelObj}
\alias{hlaCompareAllele}
\alias{hlaErrMsg}
\alias{hlaGDS2Geno}
\alias{hlaGeno2PED}
\alias{hlaGenoAFreq}
\alias{hlaGenoCombine}
\alias{hlaGenoLD}
\alias{hlaGenoMFreq}
\alias{hlaGenoMRate}
\alias{hlaGenoMRate_Samp}
\alias{hlaGenoSubset}
\alias{hlaGenoSubsetFlank}
\alias{hlaGenoSwitchStrand}
\alias{hlaHaplo2Geno}
\alias{hlaHaploSubset}
\alias{hlaLociInfo}
\alias{hlaMakeSNPGeno}
\alias{hlaMakeSNPHaplo}
\alias{hlaModelFiles}
\alias{hlaModelFromObj}
\alias{hlaModelToObj}
\alias{hlaOutOfBag}
\alias{hlaPredMerge}
\alias{hlaPublish}
\alias{hlaReport}
\alias{hlaSNPID}
\alias{hlaSampleAllele}
\alias{hlaSplitAllele}
\alias{hlaSubModelObj}
\alias{hlaUniqueAllele}
\alias{plot.hlaAttrBagClass}
\alias{plot.hlaAttrBagObj}
\alias{predict.hlaAttrBagClass}
\alias{print.hlaAttrBagClass}
\alias{print.hlaAttrBagObj}
\alias{.Last.lib}
\alias{.get_model_path}
\alias{summary.hlaAlleleClass}
\alias{summary.hlaAttrBagClass}
\alias{summary.hlaAttrBagObj}
\alias{summary.hlaSNPGenoClass}
\alias{summary.hlaSNPHaploClass}
\title{Internal PONG2 Functions}
\description{
  Internal functions used by PONG2 for KIR genotype imputation.
  These functions are adapted from the HIBAG package
  (Zheng et al. 2014, <doi:10.1016/j.ajhg.2013.12.015>)
  and are not intended to be called directly by users.
}
\value{No return value, called for side effects or internal use only.}
\keyword{internal}
