% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PONG.r
\name{kirParallelAttrBagging}
\alias{kirParallelAttrBagging}
\title{Train KIR prediction models in parallel}
\usage{
kirParallelAttrBagging(
  cl,
  hla,
  snp,
  auto.save = "",
  nclassifier = 100,
  mtry = c("sqrt", "all", "one"),
  prune = TRUE,
  rm.na = TRUE,
  stop.cluster = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{cl}{cluster object for parallel computation}

\item{hla}{KIR allele table for training}

\item{snp}{SNP genotype object}

\item{auto.save}{file path for auto saving classifiers during training}

\item{nclassifier}{number of ensemble classifiers}

\item{mtry}{number of SNPs randomly selected}

\item{prune}{logical prune classifiers}

\item{rm.na}{logical remove missing values}

\item{stop.cluster}{logical stop cluster after training completes}

\item{verbose}{logical verbose output}
}
\value{
a PONG2 model object
}
\description{
Train KIR prediction models in parallel
}
