% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PONG.r
\name{kirPredict}
\alias{kirPredict}
\title{Predict KIR genotypes from SNP data}
\usage{
kirPredict(
  object,
  snp,
  cl = FALSE,
  type = c("response+dosage", "response", "prob", "response+prob"),
  vote = c("prob", "majority"),
  allele.check = TRUE,
  match.type = c("Position", "Pos+Allele", "RefSNP+Position", "RefSNP"),
  same.strand = FALSE,
  verbose = TRUE,
  verbose.match = TRUE
)
}
\arguments{
\item{object}{a PONG2 model object}

\item{snp}{SNP genotype object}

\item{cl}{cluster object for parallel computation}

\item{type}{type of prediction}

\item{vote}{voting method}

\item{allele.check}{check alleles}

\item{match.type}{SNP matching type}

\item{same.strand}{same strand flag}

\item{verbose}{verbose output}

\item{verbose.match}{verbose matching output}
}
\value{
prediction results
}
\description{
Predict KIR genotypes from SNP data
}
